requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
anthropic>=0.18.0
pytz>=2024.1
kaleido>=0.2.1
google-analytics-data>=0.18.0
gunicorn>=21.0.0
requests>=2.31.0
